Set up simple AsyncTasks in Boat along with callbacks to run on UI thread.
Simple enough, gives effective user feedback when the connection is timing out, and fully asynchronous.
Relying on results of core comms functions as the timeout mechanism.

This is still a "best effort" system. If the button you pressed results in a time out, all it will do is tell you that.
The user must repeat the action if they want the behavior.