refactored threading into INTCATCH/unstable (#58)
* TABLE TESTING ONLY. Major refactor, shift core library comms threading into Boat object.

This should make TeleOpPanel much more simple.

Note that this is not in a working state. Only for table testing new code concepts.

* Added a while loop with a yield to force the Boat.isConnected callable to wait for core library to return.

Has a little impact on GUI responsiveness. I'm not happy about that, but I'm not sure how else to accomplish this.
A Thread.sleep() call causes the entire GUI to shutdown for the length of that call.
The while loop yield at least releases so the GUI can update.

* In process of moving more TeleOpPanel code into Boat class.

* Added timeouts to the Boat Callables with while loops in them.

* Moved marker updates into the pose listener.

* Tried out AsyncTask and callbacks. Very response but does not simplify anything.

* Moving back toward simple. If there has to be an ugly nested Runnable, it is better than crazy callbacks and AsyncTasks.

* Moving waypoint status, pose, and sensor listeners into Boat. Removed registry service option.

* Set up simple AsyncTasks in Boat along with callbacks to run on UI thread.

Simple enough, gives effective user feedback when the connection is timing out, and fully asynchronous.
Relying on results of core comms functions as the timeout mechanism.

This is still a "best effort" system. If the button you pressed results in a time out, all it will do is tell you that.
The user must repeat the action if they want the behavior.

* Added some thread safety to the pose, sensor, and waypoint state listeners.

* Added the set home and go home functions back in. Removed TabletLogger stuff.

* Tiny cleanup.